Adding support for verifying catalog signed files on windows. (#2)
On windows many operating system binaries are catalog signed, i.e. the
signature information for those binaries resides in the windows catalog
database instead of being embedded in the binaries itself. To be able to
to verify signature for such files need to find the catalog file
containing that information (based on PE file hash) and then pass it to
WinVerifyTrust to validate the signature information.